Use the silent DM feature
Make the bot less obnoxious by sending the DMs to senders and recipients
of tacos as a silent message.
This is a new Discord feature that allows the bot to increase the
notification counter but supress the notification sounds and visuals.

Also update discord.py to version 2.2 to support this feature.